daemon: extract public socket admission helpers ( W4 )
