+++ packages/nixlingd/src/lib.rs
@@ -23,8 +23,8 @@ use async_trait::async_trait;
 use nix::cmsg_space;
 use nix::fcntl::{FcntlArg, FdFlag, Flock, FlockArg, fcntl};
 use nix::sys::socket::{
-    AddressFamily, ControlMessageOwned, MsgFlags, SockFlag, SockType, UnixAddr, connect,
-    getsockopt, recv, recvmsg, send, sendto, socket, sockopt::PeerCredentials,
+    AddressFamily, ControlMessageOwned, MsgFlags, SockFlag, SockType, UnixAddr, connect, recv,
+    recvmsg, send, sendto, socket,
 };
 use nix::unistd::{self, Gid, Group, Uid, User};
 use nixling_constellation_core::TargetName;
@@ -95,8 +95,8 @@ use socket2::{Domain, SockAddr, Socket, Type};
 use supervisor::pidfd_table::{
     BrokerReapLog, PidfdEntry, PidfdRegistration, PidfdTable, PidfdTableError, WaitTermination,
 };
-use uzers::{get_user_by_uid, get_user_groups};
 
+mod admission;
 pub mod exec_detached;
 pub mod exec_session;
 pub mod exec_session_real;
@@ -107,6 +107,12 @@ pub mod supervisor;
 pub mod terminal_session;
 pub mod typed_error;
 pub mod wire;
+use admission::{
+    PeerIdentity, PeerRole, authorize_peer, gateway_display_op_requires_admin,
+    gateway_display_peer_principal, gateway_display_peer_principal_string, verb_requires_admin,
+};
+#[cfg(test)]
+use admission::{PeerOverride, TEST_PEER_OVERRIDE, TEST_PEER_OVERRIDE_LOCK};
 // `[pending restart]` machinery. Pure module + filesystem reader trait
 // so the CLI can compute the daemon-level pending-restart signal
 // post-restart without requiring /run live.
@@ -463,18 +469,6 @@ struct RuntimeIdentity {
 }
 
 #[derive(Debug, Clone)]
-struct PeerIdentity {
-    role: PeerRole,
-    uid: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PeerRole {
-    Launcher,
-    Admin,
-}
-
-#[derive(Debug, Clone)]
 struct ServerState {
     config: DaemonConfig,
     daemon_uid: u32,
@@ -563,14 +557,6 @@ fn new_gateway_display_runtime_for_tests() -> Arc<GatewayDisplayRuntime> {
         lifecycle: Box::new(DaemonGatewayLifecycle),
         preflight: None,
     })
-}
-
-#[cfg_attr(test, derive(Clone))]
-struct PeerOverride {
-    uid: u32,
-    gid: u32,
-    username: Option<String>,
-    groups: Option<Vec<String>>,
 }
 
 fn default_server_version() -> String {
@@ -2533,122 +2519,6 @@ fn run_gateway_display_owner(
             .unwrap_or_else(|_| json!({ "type": "error" })),
     };
     let _ = write_json_frame(&stream, &response);
-}
-
-fn authorize_peer(stream: &Socket, state: &ServerState) -> Result<PeerIdentity, TypedError> {
-    // Peer-identity resolution order:
-    //   1. the `#[cfg(test)]` in-process injection slot (lib unit tests that
-    //      drive `handle_connection` directly),
-    //   2. the `NIXLINGD_TEST_PEER_*` env vars (integration tests that spawn
-    //      the real daemon binary and pass them via `Command::env`; reading
-    //      env is safe under edition 2024),
-    //   3. the real `SO_PEERCRED` of the connected socket (production).
-    let peer_override = match peer_override_injected() {
-        Some(peer) => peer,
-        None => match peer_override_from_env()? {
-            Some(peer) => peer,
-            None => {
-                let peer =
-                    getsockopt(stream, PeerCredentials).map_err(io_wrap("read SO_PEERCRED"))?;
-                PeerOverride {
-                    uid: peer.uid() as u32,
-                    gid: peer.gid() as u32,
-                    username: None,
-                    groups: None,
-                }
-            }
-        },
-    };
-    let uid = peer_override.uid;
-    let _gid = peer_override.gid;
-    let username = peer_override
-        .username
-        .or_else(|| get_user_by_uid(uid).map(|user| user.name().to_string_lossy().into_owned()));
-    let _supplementary_groups = if let Some(groups) = peer_override.groups {
-        groups
-    } else if let Some(user) = get_user_by_uid(uid) {
-        get_user_groups(user.name(), user.primary_group_id())
-            .into_iter()
-            .flatten()
-            .map(|group| group.name().to_string_lossy().into_owned())
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    if uid == state.daemon_uid {
-        return Err(TypedError::AuthzNotALauncher { peer_uid: uid });
-    }
-
-    let is_launcher = username
-        .as_ref()
-        .map(|name| {
-            state
-                .config
-                .launcher_users
-                .iter()
-                .any(|launcher| launcher == name)
-        })
-        .unwrap_or(false);
-    if !is_launcher {
-        return Err(TypedError::AuthzNotALauncher { peer_uid: uid });
-    }
-
-    let role = if username
-        .as_ref()
-        .map(|name| state.config.admin_users.iter().any(|admin| admin == name))
-        .unwrap_or(false)
-    {
-        PeerRole::Admin
-    } else {
-        PeerRole::Launcher
-    };
-
-    Ok(PeerIdentity { role, uid })
-}
-
-fn verb_requires_admin(verb: &str) -> bool {
-    matches!(
-        verb,
-        "vmStart"
-            | "vmStop"
-            | "vmRestart"
-            | "switch"
-            | "boot"
-            | "test"
-            | "rollback"
-            | "gc"
-            | "keysRotate"
-            | "trust"
-            | "rotateKnownHost"
-            | "usbipBind"
-            | "usbipUnbind"
-            | "storeVerify"
-            | "migrate"
-            | "hostPrepare"
-            | "hostDestroy"
-            | "hostInstall"
-            | "hostReconcile"
-            | "readGuestConfig"
-            | "exec"
-            | "shell"
-    )
-}
-
-fn gateway_display_op_requires_admin(op: &public_wire::GatewayDisplayOp) -> bool {
-    matches!(
-        op,
-        public_wire::GatewayDisplayOp::Start(_) | public_wire::GatewayDisplayOp::Stop(_)
-    )
-}
-
-fn gateway_display_peer_principal(peer: &PeerIdentity) -> nixling_constellation_core::PrincipalId {
-    nixling_constellation_core::PrincipalId::parse(format!("uid-{}", peer.uid))
-        .expect("trusted display principal derived from numeric uid is valid")
-}
-
-fn gateway_display_peer_principal_string(peer: &PeerIdentity) -> String {
-    gateway_display_peer_principal(peer).to_string()
 }
 
 fn dispatch_request(
@@ -15732,84 +15602,6 @@ fn close_received_fds(fds: &[RawFd]) {
     for fd in fds {
         let _ = unistd::close(*fd);
     }
-}
-
-/// Test-only peer-credential injection. The accept path
-/// ([`authorize_peer`]) reads the connecting peer's identity from
-/// `SO_PEERCRED`; the accept-loop tests need to drive `handle_connection`
-/// over an in-process socketpair while pretending the peer is a specific
-/// launcher/admin uid. Rather than mutate process-global env (which is
-/// `unsafe` under edition 2024) this is injected through a `#[cfg(test)]`
-/// `Mutex`. In non-test builds it is compiled out and always `None`, so the
-/// production accept path has no test backdoor at all.
-#[cfg(test)]
-static TEST_PEER_OVERRIDE: std::sync::Mutex<Option<PeerOverride>> = std::sync::Mutex::new(None);
-
-/// Serializes the accept-loop tests that inject a [`PeerOverride`] so two of
-/// them cannot interleave on the process-global injection slot.
-#[cfg(test)]
-static TEST_PEER_OVERRIDE_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-
-#[cfg(test)]
-fn peer_override_injected() -> Option<PeerOverride> {
-    TEST_PEER_OVERRIDE
-        .lock()
-        .unwrap_or_else(|p| p.into_inner())
-        .clone()
-}
-
-#[cfg(not(test))]
-fn peer_override_injected() -> Option<PeerOverride> {
-    None
-}
-
-/// Read a peer-credential override from the `NIXLINGD_TEST_PEER_*` env vars.
-/// Used by integration tests that spawn the real daemon binary and pass these
-/// via `Command::env`; reading env is safe under edition 2024. Returns `None`
-/// (the normal production case) when `NIXLINGD_TEST_PEER_UID` is unset.
-fn peer_override_from_env() -> Result<Option<PeerOverride>, TypedError> {
-    let uid = match std::env::var("NIXLINGD_TEST_PEER_UID") {
-        Ok(value) => value
-            .parse::<u32>()
-            .map_err(|err| TypedError::InternalConfig {
-                detail: format!("NIXLINGD_TEST_PEER_UID: {err}"),
-            })?,
-        Err(std::env::VarError::NotPresent) => return Ok(None),
-        Err(err) => {
-            return Err(TypedError::InternalConfig {
-                detail: format!("NIXLINGD_TEST_PEER_UID: {err}"),
-            });
-        }
-    };
-    let gid = match std::env::var("NIXLINGD_TEST_PEER_GID") {
-        Ok(value) => value
-            .parse::<u32>()
-            .map_err(|err| TypedError::InternalConfig {
-                detail: format!("NIXLINGD_TEST_PEER_GID: {err}"),
-            })?,
-        Err(std::env::VarError::NotPresent) => uid,
-        Err(err) => {
-            return Err(TypedError::InternalConfig {
-                detail: format!("NIXLINGD_TEST_PEER_GID: {err}"),
-            });
-        }
-    };
-    let username = std::env::var("NIXLINGD_TEST_PEER_USERNAME").ok();
-    let groups = std::env::var("NIXLINGD_TEST_PEER_GROUPS")
-        .ok()
-        .map(|value| {
-            value
-                .split(',')
-                .filter(|part| !part.is_empty())
-                .map(|part| part.to_owned())
-                .collect::<Vec<_>>()
-        });
-    Ok(Some(PeerOverride {
-        uid,
-        gid,
-        username,
-        groups,
-    }))
 }
 
 fn io_wrap(context: &'static str) -> impl FnOnce(nix::errno::Errno) -> TypedError {
